Dequeue XioMsg on send-fail
If a message send hard fails, don't omit to remove it from the
send_q--this results in an assert when the queue is in safe-mode,
but with safe mode disabled, send_q would be corrupted.

Dont fall through and erase the iterator twice.  Continue the loop,
as in the incoming release case.

Signed-off-by: Matt Benjamin <matt@cohortfs.com>