XioMessenger: move header init to send epilogue
Move the late header encoding from portal thread context,
to the epilogue of _send_message_impl.

The motivation is to minimize work performed in portal thread
context, to avoid starving or delaying Accelio work.

This requires serializing the connection as in Raju's original
bugfix, but, the critcal section can be much smaller.

Signed-off-by: Matt Benjamin <mbenjamin@redhat.com>