xio: merge connect startup negotiation messages
The prototypes for these messages was merged, #if'd out, with
control flow changes in Dec. 2014.

Because get_cluster_protocol() method was identical in all the
current Messenger implementations but declared pure virtual in
Messenger, it has been moved with a default setter method to
SimplePolicyMessenger and removed from the descendants (the
setters are not removed, but call the parent's implementation,
as they have assertions about connection state).

Nb. global_seq is still essentially duplicated, but uses atomics
in XioMessenger.  Also still 32-bit (revisit?).

Signed-off-by: Matt Benjamin <mbenjamin@redhat.com>