xio: prototype support for delivery policy
Prototype support for Messenger delivery policies has been added,
building on Accelio reliable delivery primitives.

If the connection policy is Lossy, outgoing messages can be retired
immediately (i.e., in on_ow_msg_send_complete).  Otherwise, we request
a delivery receipt, and retire the message in on_msg_delivered.

In the event of disconnection, any messages queued on Accelio but not
delivered will be returned in on_msg_error.  If the connection policy
is Lossless, we now queue such messages for re-delivery at the next
reconnect.

Signed-off-by: Matt Benjamin <mbenjamin@redhat.com>