xio: remove broken Accelio session interaction w/startup
An attempt to mix Accelio's session startup negotiation w/the Ceph
message style was unsatisfactory.

The idea to use Accelio's was a late addition proposed by Eyal, and
in hindsight, doesn't appear to have a notable advantage over the
message approach--in particular, it would essentially require a
separate encoded startup-message grammar (like the current Ceph TCP
startup protocol), which we are trying to avoid.

In place of serialized data, put a standard CEPH_BANNER string in the
private data.

Signed-off-by: Matt Benjamin <mbenjamin@redhat.com>