xio: expose native layout of xio_msg_ex structures
Adds a pointer to the underlying xio_msg_ex structure to the
xio_msg_buffer raw subclass.  In addition, adds a method to
introspect buffer::ptr for an underlying xio_msg_ex, if it is
present.

Ceph code such as erasure code offload can require access to
details of the underlying message, such as any handle to
registered memory.

Signed-off-by: Matt Benjamin <mbenjamin@redhat.com>