client: add delegation support for cephfs
Add the ability for ceph userland clients to request a delegation for a
cephfs open file. With this, userland will get a callback if there is a
request for access that conflicts with the delegation. The client is
then expected to return the delegation at which point the conflicting
access will succeed.

Signed-off-by: Jeff Layton <jlayton@redhat.com>