rgwlc:  unify rgwlc and inventory attribute processing
The idea is to permit these attributes to share a reliable mapping
to lifecycle's work list, so that the same work cycle can service
lifecycle and inventory processing.

As a side effect, integrate the corresponding bucket attr updates
into retry_raced_bucket_write(...), which seemed to be implied but
not present.

set_bucket_config(...) failing...but..?

Signed-off-by: Matt Benjamin <mbenjamin@redhat.com>