rgw: improved admin token diagnostics.
As a result of previous bugs, radosgw was caught trying to use
a defective cached admin token.  This fix includes logic to prevent this;
if a token is bad, it should not be cached, and should generate
a more obvious error message.

This fix:
/1/ be far pickier about incoming admin tokens.
/2/ try to produce more useful error messaging upon misbehavior or misconfiguration.

Signed-off-by: Marcus Watts <mwatts@redhat.com>
Resolves: rhbz#2099828
Resolves: rhbz#2116926
(cherry picked from commit 0bdc6db9a80af40dd496b05674a938d406a9f6f5)

Conflicts:
	src/rgw/rgw_auth_keystone.cc