rgw/reshard: stop recording logs when reshard failed
A log will be recorded together with index entry in the logrecord
state, ant it should be stopped when reshard fails. Add a judge
every certain time at the logrecord state, the reshard has been
failed if the reshard_lock can be held, then clear the state and
stop recording logs.

Signed-off-by: Mingyuan Liang <liangmingyuan@baidu.com>
Signed-off-by: Matt Benjamin <mbenjamin@redhat.com>