rgw/reshard: Backward Compatibility
The privious release only has one reshard phase: the progress phase
which will block client writes. Because our release contains this
phase and the process is same too, that means it is superset of
privious release. So when privious rgw initiates a reshard, it will
execute as before.
When a updated rgw initiates a reshard, it firstly enter the
logrecord phase which privious releases do not realized. That means
the nodes which do not upgraded will deal with client write
operations without recording logs. It may leads to part of these
index entries missed. So we forbit this scene by controlling the
cls_rgw_set_bucket_resharding_op version, then the updated rgw will
stop resharding when meeting privious releases nodes.

Signed-off-by: Mingyuan Liang <liangmingyuan@baidu.com>